+++ apps/docs/src/app/pages/docs/components/charts/line-area-chart/line-area-chart.component.ts
@@ -30,13 +30,18 @@ export class LineAreaChartComponent {
       level: JumpNavLevel.h3
     },
     {
+      key: 'Events',
+      value: 'events',
+      level: JumpNavLevel.h2
+    },
+    {
       key: 'Contract',
       value: 'contract',
       level: JumpNavLevel.h2
     }
   ] as IJumpNav[];
-  markupCode = `<pcac-line-area-chart [config]="config"></pcac-line-area-chart>`;
+  markupCode = `<pcac-line-area-chart [config]="config" (dotClicked)="onClicked($event)"></pcac-line-area-chart>`;
   importCode = `import { PcacLineAreaChartModule } from '@pioneer-code/pioneer-charts';`;
-  constructor(public pcacService: PcService) { }
+  constructor(public pcService: PcService) { }
 }