+++ projects/pcac/src/lib/pie-chart/pie-chart.builder.ts
@@ -3,17 +3,28 @@ import { Injectable, ElementRef } from '@angular/core';
 /**
  * D3
  */
-import { arc, pie, DefaultArcObject, Arc, Pie } from 'd3-shape';
+import { arc, pie, DefaultArcObject, Arc, Pie, PieArcDatum } from 'd3-shape';
 import { select } from 'd3-selection';
 import { interpolate } from 'd3-interpolate';
 import { transition } from 'd3-transition';
+import { color } from 'd3-color';
 
 /**
  * Lib
  */
 import { IPcacPieChartConfig } from './pie-chart.model';
-import { PcacChart } from '../core/chart';
-import { IPcacData } from '../core/chart.model';
+import {
+  PcacChart,
+  IPcacData,
+  PcacAxisBuilder,
+  PcacGridBuilder,
+  PcacColorService,
+  PcacTooltipBuilder,
+  PcacTransitionService
+} from '../core';
+
+import { Observable } from 'rxjs/internal/Observable';
+import { Subject } from 'rxjs/internal/Subject';
 
 export interface IPieChartBuilder {
   buildChart(chartElm: ElementRef, config: IPcacPieChartConfig): void;
@@ -25,6 +36,24 @@ export class PieChartBuilder extends PcacChart implements IPieChartBuilder {
   private arcShape: Arc<any, DefaultArcObject>;
   private arcOverShape: Arc<any, DefaultArcObject>;
   private pieAngles: Pie<any, number | {}>;
+  private sliceClickedSource = new Subject<IPcacData>();
+  sliceClicked$ = this.sliceClickedSource.asObservable();
+
+  constructor(
+    public axisBuilder: PcacAxisBuilder,
+    public gridBuilder: PcacGridBuilder,
+    public transitionService: PcacTransitionService,
+    public tooltipBuilder: PcacTooltipBuilder,
+    public colorService: PcacColorService
+  ) {
+    super(
+      axisBuilder,
+      gridBuilder,
+      transitionService,
+      tooltipBuilder,
+      colorService
+    );
+  }
 
   buildChart(chartElm: ElementRef, config: IPcacPieChartConfig): void {
     this.initializeChartState(chartElm, config);
@@ -57,26 +86,31 @@ export class PieChartBuilder extends PcacChart implements IPieChartBuilder {
       .enter().append('g')
       .attr('class', 'pcac-arc')
       .append('path')
-      .style('fill', (d: any, i: number) => {  // TODO: Strongly type
+      .style('fill', (d: PieArcDatum<IPcacData>, i: number) => {
         return this.colors[i];
       })
-      .on('mouseover', function (d: any) {  // TODO: Strongly type
+      .on('mouseover', function (d: PieArcDatum<IPcacData>, i: number) {
         select(this).transition(transition()
           .duration(self.transitionService.getTransitionDuration() / 3))
-          .attr('d', self.arcOverShape);
+          .attr('d', self.arcOverShape)
+          .style('fill', color(self.colors[i]).darker(1).toString())
       })
-      .on('mousemove', function (d: any) {  // TODO: Strongly type
-        self.tooltipBuilder.showBarTooltip(d);
+      .on('mousemove', function (d: PieArcDatum<IPcacData>) {
+        self.tooltipBuilder.showBarTooltip(d.data);
       })
-      .on('mouseout', function () {
+      .on('mouseout', function (d: PieArcDatum<IPcacData>, i: number) {
         self.tooltipBuilder.hideTooltip();
         select(this).transition(transition()
           .duration(self.transitionService.getTransitionDuration() / 3))
-          .attr('d', self.arcShape);
+          .attr('d', self.arcShape)
+          .style('fill', self.colors[i]);
+      })
+      .on('click', (d: PieArcDatum<IPcacData>, i: number) => {
+        this.sliceClickedSource.next(d.data);
       })
       .transition(transition()
         .duration(this.transitionService.getTransitionDuration()))
-      .attrTween('d', (b: any) => {  // TODO: Strongly type
+      .attrTween('d', (b: PieArcDatum<IPcacData>) => {
         return this.tweenChart(b);
       });
   }